+++ framework/proto/flwr/proto/clientappio.proto
@@ -77,6 +77,9 @@ service ClientAppIo {
   rpc ConfirmMessageReceived(ConfirmMessageReceivedRequest)
       returns (ConfirmMessageReceivedResponse) {}
 
+  // Create a task
+  rpc CreateTask(CreateTaskRequest) returns (CreateTaskResponse) {}
+
   /////////////////////////////////////////////////////////////////////////////
   // Specific endpoints for ClientAppIo
   /////////////////////////////////////////////////////////////////////////////
