+++ framework/py/flwr/supercore/auth/__init__.py
@@ -16,7 +16,6 @@
 
 
 from .policy import (
-    APPIO_METHOD_AUTH_POLICY,
     CLIENTAPPIO_METHOD_AUTH_POLICY,
     SERVERAPPIO_METHOD_AUTH_POLICY,
     MethodTokenPolicy,
@@ -30,7 +29,6 @@ from .superexec import (
 from .superexec_secret import add_superexec_auth_secret_args, load_superexec_auth_secret
 
 __all__ = [
-    "APPIO_METHOD_AUTH_POLICY",
     "CLIENTAPPIO_METHOD_AUTH_POLICY",
     "MethodTokenPolicy",
     "SERVERAPPIO_METHOD_AUTH_POLICY",
