+++ framework/py/flwr/supercore/interceptors/appio_token_interceptor_test.py
@@ -29,13 +29,11 @@ from flwr.proto.appio_pb2 import (  # pylint: disable=E0611
     PushAppMessagesRequest,
     PushAppOutputsRequest,
 )
-from flwr.proto.appio_pb2_grpc import AppIoServicer as GrpcAppIoServicer
 from flwr.proto.clientappio_pb2_grpc import ClientAppIoServicer
 from flwr.proto.message_pb2 import PushObjectRequest  # pylint: disable=E0611
 from flwr.proto.serverappio_pb2 import GetNodesRequest  # pylint: disable=E0611
 from flwr.proto.serverappio_pb2_grpc import ServerAppIoServicer
 from flwr.supercore.auth import (
-    APPIO_METHOD_AUTH_POLICY,
     CLIENTAPPIO_METHOD_AUTH_POLICY,
     SERVERAPPIO_METHOD_AUTH_POLICY,
 )
@@ -408,14 +406,6 @@ class TestMethodPolicyMaps(TestCase):
     }
 
     @staticmethod
-    def _appio_rpc_methods() -> set[str]:
-        return {
-            f"/flwr.proto.AppIo/{name}"
-            for name, ref in inspect.getmembers(GrpcAppIoServicer)
-            if inspect.isfunction(ref) and not name.startswith("_")
-        }
-
-    @staticmethod
     def _serverappio_rpc_methods() -> set[str]:
         return {
             f"/flwr.proto.ServerAppIo/{name}"
@@ -435,20 +425,6 @@ class TestMethodPolicyMaps(TestCase):
         """ServerAppIo policy map should cover all RPC methods exactly."""
         expected_methods = self._serverappio_rpc_methods()
         self.assertEqual(set(SERVERAPPIO_METHOD_AUTH_POLICY), expected_methods)
-
-    def test_appio_policy_has_full_coverage(self) -> None:
-        """AppIo policy map should cover all AppIo RPC methods exactly."""
-        expected_methods = self._appio_rpc_methods()
-        self.assertEqual(set(APPIO_METHOD_AUTH_POLICY), expected_methods)
-
-    def test_appio_methods_require_auth(self) -> None:
-        """AppIo methods should all require App token auth."""
-        no_auth_methods = {
-            method.rsplit("/", maxsplit=1)[-1]
-            for method, policy in APPIO_METHOD_AUTH_POLICY.items()
-            if not policy.requires_token
-        }
-        self.assertEqual(no_auth_methods, set())
 
     def test_only_expected_no_auth_methods_exist(self) -> None:
         """Only bootstrap methods should be marked no-auth in the policy table."""
@@ -484,7 +460,23 @@ class TestFactoryFunctions(TestCase):
         intercepted = interceptor.intercept_service(
             lambda _: _make_unary_handler(),
             _HandlerCallDetails(
-                "/flwr.proto.AppIo/CreateTask",
+                "/flwr.proto.ServerAppIo/CreateTask",
+                invocation_metadata=((APP_TOKEN_HEADER, "valid-token"),),
+            ),
+        )
+
+        response = cast(str, intercepted.unary_unary(GetNodesRequest(run_id=1), Mock()))
+        self.assertEqual(response, "ok")
+
+    def test_clientappio_factory_uses_appio_create_task_policy(self) -> None:
+        """ClientAppIo factory should also allow the shared CreateTask path."""
+        state = _TokenState({"valid-token": 1})
+        interceptor = create_clientappio_token_auth_server_interceptor(lambda: state)
+
+        intercepted = interceptor.intercept_service(
+            lambda _: _make_unary_handler(),
+            _HandlerCallDetails(
+                "/flwr.proto.ClientAppIo/CreateTask",
                 invocation_metadata=((APP_TOKEN_HEADER, "valid-token"),),
             ),
         )
